add the ability to add from settings
